% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_DA_fit.R
\name{boot_DA_fit}
\alias{boot_DA_fit}
\title{Boot_DA_fit}
\usage{
boot_DA_fit(dat, start_val_grid, n)
}
\arguments{
\item{dat}{Input data}

\item{start_val_grid}{Gird with the start values for a four parameter model (ct, CRbreak, S2, tf)}

\item{n}{Number of bootstrap samples}
}
\value{

}
\description{
Boot_DA_fit
}
\examples{
boot_DA_fit(dat, start_val_grid, n)
}
