% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistart_fit_p4_model.R
\name{multistart_fit_p4_model}
\alias{multistart_fit_p4_model}
\title{Fit four parameter models with multiple start parameters}
\usage{
multistart_fit_p4_model(data, start_val_grid)
}
\arguments{
\item{data}{Dark adaptation data with a column 'Time' and 'Threshold'}

\item{start_val_grid}{Gird with the start values for a four parameter model (ct, CRbreak, S2, tf)}
}
\value{

}
\description{
This function allows you to fit four parameter models using multiple start parameters
}
\examples{
multistart_fit_p4_model(data, start_val_grid)
}
