% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_DA_fit.R
\name{select_best_DA_fit}
\alias{select_best_DA_fit}
\title{Select best model based on AIC}
\usage{
select_best_DA_fit(candidate_fits)
}
\arguments{
\item{candidate_fits}{Multiple model fits obtained by the multistart_fit_p4_model() function}
}
\value{

}
\description{
This function allows you to select the best model based on AIC
}
\examples{
select_best_DA_fit(candidate_fits)
}
